feat[Service]: Add function to validate access token
