+++ src/poly/services/__init__.py
@@ -1,0 +1,3 @@
+from fastapi.security import OAuth2PasswordBearer
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")
